admin signup

+++ controllers/commentcontroller.js
@@ -33,7 +33,7 @@ router.get('/all-comments', validateSession, (req, res) => {
 //Edit comment
 router.put('./:id', validateSession, (req, res) => {
     if (!req.errors) {
-        CommentModel.update(req.body.post, {where: {owner: req.user.id, id: req.params.id}})
+        CommentModel.update(req.body.post, {where: {owner: req.user.id, id: req.params.id}} || {where: {admin: req.user.admin.true, id: req.params.id}})
         .then(data => res.status(200).json(data))
         .catch(err => res.status(500).json(err));
     } else {
@@ -44,7 +44,7 @@ router.put('./:id', validateSession, (req, res) => {
 //Delete comment 
 router.delete('./:id', validateSession, (req, res) => {
     if (!req.errors){
-        CommentModel.destroy({where: { owner: req.user.id, id: req.params.id}})
+        CommentModel.destroy({where: { owner: req.user.id, id: req.params.id}} || {where: {admin: req.user.admin.true, id: req.params.id}})
         .then(data => res.status(200).json(data))
         .catch(err => res.status(500).json(err));
     } else {
