+++ controllers/postcontroller.js
@@ -59,7 +59,7 @@ router.put('./:id', validateSession, (req, res) => {
 //Delete post 
 router.delete('./:id', validateSession, (req, res) => {
     if (!req.errors){
-        PostModel.destroy({where: { owner: req.user.id, id: req.params.id}})
+        PostModel.destroy({where: {owner: req.user.id, id: req.params.id}} || {where: {admin: req.user.admin.true, id:req.params.id}})
         .then(data => res.status(200).json(data))
         .catch(err => res.status(500).json(err));
     } else {
@@ -69,3 +69,9 @@ router.delete('./:id', validateSession, (req, res) => {
 
 
 module.exports = router
+
+// this is the non-admin token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MiwiaWF0IjoxNTk0NzQ5NDU5LCJleHAiOjE1OTQ4MzU4NTl9.CZxejNdKq7xPn1mROtgQ-DkhGqLwDH2HE-nlnjCPgl4
+
+// this is the admin token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MywiaWF0IjoxNTk0NzQ5Nzc1LCJleHAiOjE1OTQ4MzYxNzV9.OmTQcELiZj0V8v1aa9S3T0VWQHrxw84bynm978Th1R8
+//username: admin
+//password: adminpassword
